+++ ui/utils.go
@@ -8,6 +8,7 @@ import (
 	"github.com/RocketChat/Rocket.Chat.Go.SDK/models"
 )
 
+// Return first letter of a string.
 func getStringFirstLetter(str string) string {
 	letter := "U"
 	if len(str) > 0 {
@@ -16,6 +17,8 @@ func getStringFirstLetter(str string) string {
 	return letter
 }
 
+// Compare today's date time with the token expiration date to validate token.
+// Take token expiration time as string in argument and then converted into value of type time.Time to compare with today's date time.
 func CheckForTokenExpiration(tokenExpirationTime string) bool {
 	today := time.Now()
 	i, _ := strconv.ParseInt(tokenExpirationTime, 10, 64)
@@ -23,22 +26,21 @@ func CheckForTokenExpiration(tokenExpirationTime string) bool {
 	return today.Before(tokenExpires)
 }
 
+// To extract sub string from a particular position in a string until the string terminates or there is a space in the string.
 func stringUsernameExtractor(str string, currentPos int) string {
 	length := len(str)
 	chars := []rune(str)
 	word := ""
 	pos := currentPos
-	for pos < length {
-		if string(chars[pos]) != "" && string(chars[pos]) != " " {
-			word += string(chars[pos])
-		} else {
-			break
-		}
+	for pos < length && string(chars[pos]) != "" && string(chars[pos]) != " " {
+		word += string(chars[pos])
 		pos++
 	}
 	return word
 }
 
+// When user selects a username from the list then add that complete username in place of the username first letters typed by user.
+// Return complete string after adding username at the starting position of the username first letters typed by user.
 func usernameAutoCompleteString(str string, replacement string, index int, jump int) string {
 	return str[:index] + replacement + str[index+jump:]
 }
